Refactor dialogs, converters, filtering, and registry services
Major codebase refactor:
- Modularized dialog, registry, and JSON services with new interface-driven and type-safe implementations.
- Replaced all C2*Converter classes with new, generic converters and updated XAML bindings.
- Enhanced filtering with min/max level support and improved FiltersPanel UI.
- Switched selection logic to use arrays and added status bar selection count.
- Centralized app-wide constants in App.Constants.
- Updated undo/redo command structure and naming.
- Changed song IDs to BigInteger for flexibility.
- Cleaned up obsolete classes and updated resource files for new UI features.

+++ C2SR/ViewModels/C2SongViewModel.cs
@@ -2,6 +2,7 @@
 using C2SR.Models;
 using C2SR.Services;
 using CommunityToolkit.Mvvm.ComponentModel;
+using System.Numerics;
 
 namespace C2SR.ViewModels
 {
@@ -17,7 +18,7 @@ namespace C2SR.ViewModels
         readonly C2Song song;
 
         #region Properties
-        public long ID => song.ID;
+        public BigInteger ID => song.ID;
         public string Name => song.Name;
         public string Artist => song.Artist;
         public decimal Bpm => song.Bpm;
